ok prolog refactor
